third_party/blink/web_tests/svg/dynamic-updates/SVGMaskElement-svgdom-maskUnits-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'maskUnits' property of the SVGMaskElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS maskElement.maskUnits.baseVal is SVGUnitTypes.SVG_UNIT_TYPE_OBJECTBOUNDINGBOX
PASS maskElement.maskUnits.baseVal is SVGUnitTypes.SVG_UNIT_TYPE_USERSPACEONUSE
PASS successfullyParsed is true

TEST COMPLETE
